some corrections

+++ projekat/demo/src/main/java/rs/team15/repository/MenuItemRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import rs.team15.model.MenuItem;
 
-public interface MenuItemRepository extends JpaRepository <MenuItem, Integer> {
+public interface MenuItemRepository extends JpaRepository <MenuItem, Long> {
 
     Collection <MenuItem> findByMid(Integer id);
 
-    Collection <MenuItem> findByTypeAndDeletedAndMid(String type, boolean deleted, Integer id);
+    Collection <MenuItem> findByTypeAndDeletedAndMid(String type, boolean deleted, Long id);
 }
